Add movement3d module and use for avatar motion
